pyproject.toml: fix broken package due to missing NTFS submodule
FATtools now supports NTFS via a submodule, which is a first for this
project. Considering the submodule is indiscriminately imported
regardless of the type of volumes to work with, missing the submodule
means that no command works at all.

Because the package built by the build system didn't contain the NTFS
submodule, all PYPI releases between v1.1.0 and now are broken. This
fixes the issue by making sure the submodule is indeed included in the
built package.

Fixes: c9509ec58429 ("Fixed and enhanced with basic NTFS support")
Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>