pyproject.toml: migrate to PEP-639
The license field in PEP-639 is now an SPDX license expression.[1]

Note that while the gpl.txt license file text is the one of the
GPL-3.0-or-later, FATtools/__init__.py and FATtools/NTFS/__init__.py
(the only two files with a mention of GPL) only mention "GPLv3" meaning
GPL-3.0-only is expected, otherwise "GPLv3+" or "or later" or "or any
later version" should have been written. I'm not a lawyer but that is my
understanding of the license.

Because `python -m build` (used in GitHub workflow) creates a venv with
the requirements, it'll build with a recent setuptools. However, to make
sure it builds with a recent enough setuptools, hardcode the minimum
version that supports PEP-639: 77.0.3.[2]

Finally, following the recommendations of the setuptools project
itself[3], let's remove the "License ::" classifier.

[1] https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license
[2] https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#id9
[3] https://packaging.python.org/en/latest/guides/licensing-examples-and-user-scenarios/#basic-example

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>