pyproject.toml: ship GPL license text
It is usually expected from open-source software to ship the license
with the code. Build systems like the Yocto Project expects some license
files to be provided to be able to monitor changes to licenses used by
the project it's building.

My understanding is that the GPL-3.0-or-later requires you to provide
the license text with verbatim source code (which would be the sdist
package uploaded to PYPI I believe), c.f. 4. Conveying Verbatim Copies:

  You may convey verbatim copies of the Program's source code as you
  receive it, in any medium, provided that you conspicuously and
  [...] give all
  recipients a copy of this License along with the Program.

Let's add the GPL license text to the python package.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>